Final refactorization of DBus modules
* Move module base classes to pyanaconda.dbus.modules.
* Move task base classes to pyanaconda.dbus.task.
* Move kickstart specification classes to pyanaconda.core.kickstart.
* Remove some unnecessary directories.
