+++ pyanaconda/dbus/Makefile.am
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = errors
+SUBDIRS = errors module task
 
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
 dbusdir = $(pkgpyexecdir)/dbus
