+++ pyanaconda/dbus_addons/baz/baz_interface.py
@@ -17,9 +17,9 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
-from pyanaconda.dbus.interface import dbus_interface
 from pyanaconda.dbus.objects import BAZ
-from pyanaconda.modules.base_interface import KickstartModuleInterface
+from pyanaconda.dbus.interface import dbus_interface
+from pyanaconda.dbus.module import KickstartModuleInterface
 
 
 @dbus_interface(BAZ.interface_name)
