+++ pyanaconda/modules/boss/Makefile.am
@@ -15,8 +15,6 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = install_manager
-
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
 bossdir = $(pkgpyexecdir)/modules/boss
 boss_PYTHON = $(srcdir)/*.py
