+++ pyanaconda/modules/boss/boss_interface.py
@@ -17,11 +17,11 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
-
 from pyanaconda.dbus.interface import dbus_interface
 from pyanaconda.dbus.objects import BOSS, BOSS_ANACONDA
 from pyanaconda.dbus.template import InterfaceTemplate
 from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
+from pyanaconda.dbus.typing import Bool, Str, Int
 
 
 @dbus_interface(BOSS.interface_name)
