+++ pyanaconda/modules/boss/installation_interface.py
