+++ pyanaconda/modules/boss/installation_manager.py
