+++ pyanaconda/modules/timezone/timezone_interface.py
@@ -20,7 +20,7 @@
 from pyanaconda.dbus.objects import TIMEZONE
 from pyanaconda.dbus.property import emits_properties_changed
 from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
-from pyanaconda.modules.base_interface import KickstartModuleInterface
+from pyanaconda.dbus.module import KickstartModuleInterface
 from pyanaconda.dbus.interface import dbus_interface
